pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "frm_galt"
version = "0.0.1"
authors = [
  { name="John Galt", email="frm.calcs@gmail.com" },
]
description = "A python package for quantitative finance and derivative pricing"
readme = "README.md"
requires-python = ">=3.10"
dependencies=[           
    "holidays",
    "pandas",
    "numpy",
    "pandas_market_calendars",
    "datetime",
    "scipy",
    "matplotlib",
    "numba",
      ]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",  
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/frmcalcs/frm"
"Website" = "https://www.frmcalcs.com/app_frm/"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
exclude = ["references", "testing"]  # Add these lines to exclude specific packages
